Include status summary when viewing commits
Due to using gitv I like to turn off folding of commits so that I can
easily browse through the history and easily read the full diff for
small commits; but with folding disabled it no longer gives a good
overview for larger commits.  Add the --stat option when calling `git
show` to have an overview included, also need to add the --patch option
to avoid the diff being suppressed due to that.